[CF]增加宇树电机(no sdk),分离遥控器控制

+++ src/motor_control_ros2/src/config_parser.cpp
@@ -64,13 +64,13 @@ MotorConfig ConfigParser::parseMotorConfig(const YAML::Node& node) {
   
   // GO8010 电机特定参数
   if (node["gear_ratio"]) {
-    config.gear_ratio = node["gear_ratio"].as<float>();
+    config.gear_ratio = node["gear_ratio"].as<double>();
   }
   if (node["k_pos"]) {
-    config.k_pos = node["k_pos"].as<float>();
+    config.k_pos = node["k_pos"].as<double>();
   }
   if (node["k_spd"]) {
-    config.k_spd = node["k_spd"].as<float>();
+    config.k_spd = node["k_spd"].as<double>();
   }
   
   return config;
